Move `HostList` to `Icinga\Module\Icingadb\Widget\ItemList`
