+++ library/Icingadb/Widget/BaseHostListItem.php
@@ -7,6 +7,7 @@ namespace Icinga\Module\Icingadb\Widget;
 use Icinga\Module\Icingadb\Common\Links;
 use Icinga\Module\Icingadb\Common\NoSubjectLink;
 use Icinga\Module\Icingadb\Model\Host;
+use Icinga\Module\Icingadb\Widget\ItemList\HostList;
 use ipl\Html\Attributes;
 use ipl\Html\HtmlElement;
 use ipl\Html\Text;
