+++ library/Icingadb/Widget/ItemList/HostList.php
@@ -2,9 +2,13 @@
 
 /* Icinga DB Web | (c) 2020 Icinga GmbH | GPLv2 */
 
-namespace Icinga\Module\Icingadb\Widget;
+namespace Icinga\Module\Icingadb\Widget\ItemList;
 
 use Icinga\Module\Icingadb\Common\Links;
+use Icinga\Module\Icingadb\Widget\HostListItem;
+use Icinga\Module\Icingadb\Widget\HostListItemDetailed;
+use Icinga\Module\Icingadb\Widget\HostListItemMinimal;
+use Icinga\Module\Icingadb\Widget\StateList;
 use ipl\Web\Url;
 
 /**
